feat: Entity 컬럼 추가 및 수정, 모든 int => Integer로 변경

+++ backend/src/main/java/com/ssafy/d3v/backend/feedback/entity/Feedback.java
@@ -25,7 +25,7 @@ public class Feedback extends BaseEntity {
     @Id
     @Column(name = "feedback_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long feedbackId;
+    private Integer feedbackId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "answer_id", nullable = false)
